added: crop image feature to c++ code, now output panorama is rectangular; waiting proces feature with progress circle in layout; fixed: NDK incopatibility - downgraded to ndk-r15; saving image; changed area for taking pictures to smaller;

+++ android/src/main/java/study/acodexm/utils/ImageRW.java
@@ -6,11 +6,16 @@ import android.graphics.BitmapFactory;
 import android.os.Environment;
 import android.util.Log;
 
+import org.opencv.core.Mat;
+import org.opencv.imgcodecs.Imgcodecs;
+
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 public class ImageRW {
     private static final String TAG = ImageRW.class.getSimpleName();
@@ -35,6 +40,31 @@ public class ImageRW {
         } else {
             Log.d(TAG, "File saving failed");
         }
+    }
+
+    public static boolean saveResultImageExternal(Mat result) {
+        Log.d(TAG, "saveResultImageExternal: begin saving");
+        File folder = new File(Environment.getExternalStorageDirectory()
+                + "/PanoramaApp");
+        Date date = new Date();
+        SimpleDateFormat simple = new SimpleDateFormat("HH-mm-ss__dd_MM_yyyy");
+        final String fileName = folder.getAbsolutePath() + "/panorama_" +
+                simple.format(date) + ".png";
+        Log.d(TAG, "saveResultImageExternal: filename: " + fileName);
+        boolean success = true;
+        if (!folder.exists()) {
+            success = folder.mkdirs();
+        }
+        if (success) {
+            try {
+                return Imgcodecs.imwrite(fileName, result);
+            } catch (Exception e) {
+                Log.e(TAG, "File saving failed", e);
+            }
+        } else {
+            Log.d(TAG, "File saving failed");
+        }
+        return false;
     }
 
     public static void saveImageForTextureExternal(byte[] bytes, int currentPictureId,
@@ -65,7 +95,7 @@ public class ImageRW {
                 + "/PanoramaApp/temp/");
         if (dir.isDirectory()) {
             String[] children = dir.list();
-            if (children.length > 0)
+            if (children != null && children.length > 0)
                 for (String aChildren : children) {
                     if (new File(dir, aChildren).delete()) {
                         Log.d(TAG, "file " + aChildren.trim() + " deleted");
@@ -76,7 +106,7 @@ public class ImageRW {
         }
     }
 
-    public static Bitmap loadImageExternal(int currentPictureId) {
+    static Bitmap loadImageExternal(int currentPictureId) {
         final String fileName = Environment.getExternalStorageDirectory().getAbsolutePath()
                 + "/PanoramaApp/temp/" + currentPictureId + ".png";
         Bitmap bitmap = null;
@@ -90,7 +120,7 @@ public class ImageRW {
         return bitmap;
     }
 
-    private static byte[] resizeImage(byte[] bytes, int PHOTO_WIDTH, int PHOTO_HEIGHT) {
+    static byte[] resizeImage(byte[] bytes, int PHOTO_WIDTH, int PHOTO_HEIGHT) {
         Bitmap original = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
         Bitmap resized = Bitmap.createScaledBitmap(original, PHOTO_WIDTH, PHOTO_HEIGHT, false);
         ByteArrayOutputStream blob = new ByteArrayOutputStream();
